moved fixture to header file
